Error

+++ src/bootloader/stage2/stdint.h
@@ -4,10 +4,10 @@ typedef signed char int8_t;
 typedef unsigned char uint8_t;
 typedef signed short int16_t;
 typedef unsigned short uint16_t;
-typedef unsigned int uint32_t;
-typedef signed int int32_t;
-typedef unsigned long uint64_t;
-typedef signed long int64_t;
+typedef unsigned long int uint32_t;
+typedef signed long int int32_t;
+typedef unsigned long long int uint64_t;
+typedef signed long long int int64_t;
 
 typedef uint8_t bool;
